requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
numpy==1.11.0
Pillow==3.2.0
Werkzeug==0.11.9
